[FIX] cleaner code

+++ Assets/ScriptableObjects/AutoCostUpgradeTemplate.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "NewAutoCostUpgrade", menuName = "Data/Upgrade/Auto Cost")]
@@ -13,7 +11,7 @@ public class AutoCostUpgradeTemplate : UpgradeTemplate
     {
         ResourceDict newCost = new ResourceDict();
         // Formula is cost = baseCost * exponent ^ rank
-        baseCost.ForEach((key, value) => newCost.Add(key, (int)(value*Mathf.Pow(exponent, rank))));
+        baseCost.ForEach((key, value) => newCost.Add(key, (long)(value*Mathf.Pow(exponent, rank))));
         
         return newCost;
     }
